[feat] - zustand todos 토글 및 필터링

+++ 03_zustand/src/components/todo/TodoItem.tsx
@@ -7,12 +7,20 @@ interface TodoItemProps {
 
 const TodoItem = ({ todo }: TodoItemProps) => {
   const { id, title, content, deadline, isDone } = todo;
-  const { deleteTodo } = useTodoStore();
+  const { deleteTodo, toggleDoneTodo } = useTodoStore();
 
   // 삭제
   const handleDeleteButton = () => {
-    deleteTodo(id);
-    alert("삭제가 완료되었습니다.");
+    const deleteConfirm = window.confirm("정말 삭제하시겠습니까?");
+    if (deleteConfirm) {
+      deleteTodo(id);
+      alert("삭제가 완료되었습니다.");
+    }
+  };
+
+  // 토글
+  const handleDoneToggleButton = () => {
+    toggleDoneTodo(id, isDone);
   };
 
   return (
@@ -21,6 +29,9 @@ const TodoItem = ({ todo }: TodoItemProps) => {
       <p>{content}</p>
       <p>{deadline}</p>
       <button onClick={handleDeleteButton}>삭제</button>
+      <button onClick={handleDoneToggleButton}>
+        {isDone ? "할 일 완료" : "할 일 취소"}
+      </button>
     </li>
   );
 };
